Make it work on workers
