+++ packages/django-cf/django_cf/middleware/__init__.py
@@ -1,0 +1,1 @@
+from .CloudflareAccessMiddleware import CloudflareAccessMiddleware
